Split out issue commands into dedicated classes
